refactor: extract keybinding handlers from main event loop
- Create keybindings module with focused helper functions
- Replace giant match statement with organized function calls
- Group navigation, arrow, and textarea editing logic separately
- Reduce main event loop complexity by ~150 lines
